refactor: TDET-50 Uploaders refactored to be called from within the operators
